rental_project/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-appconf==1.0.6
django-compressor==4.5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-decouple==3.8
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.3
